web_scraper/requirements.txt
beautifulsoup4
requests
tenacity
pyrate-limiter
python-dotenv
pytest
